requirements.txt
python-telegram-bot==21.4
aiohttp==3.9.5
